+++ packages/nadle/src/core/utilities/index.ts
@@ -1,4 +1,4 @@
-export { NadleError } from "./nadle-error.js";
 export { MaybeArray } from "./maybe-array.js";
 export type { Callback, Awaitable, Resolver } from "./types.js";
 export { type SupportLogLevel, SupportLogLevels } from "./consola.js";
+export { NadleError, ConfigurationError, TaskNotFoundError, CyclicDependencyError, TaskExecutionError } from "./nadle-error.js";